move warn function to utils module
